#3 - Support showing duration in humanized way and update title of reminder notification
