+++ hkb_daemon/Cargo.toml
@@ -9,6 +9,7 @@ edition = "2021"
 dirs = { workspace = true }
 tokio = { workspace = true }
 diesel_migrations = { workspace = true }
+notify-rust = { version = "4.11.0" }
 hkb_date = { path = "../hkb_date", version = "0.1.0" }
 hkb_daemon_core = { path = "../hkb_daemon_core", version = "0.1.0" }
 hkb_core = { path = "../hkb_core", version = "0.1.0", features = ["sqlite-database"] }
