+++ hkb_daemon/src/main.rs
@@ -1,3 +1,4 @@
+use std::collections::HashSet;
 use std::time::Duration;
 
 use diesel_migrations::{embed_migrations, EmbeddedMigrations};
@@ -5,7 +6,8 @@ use hkb_core::database::init_database;
 use hkb_core::database::services::reminders::*;
 use hkb_core::logger::{self, debug, error, info, AppenderType};
 use hkb_daemon_core::server::Server;
-use hkb_date::date::SimpleDate;
+use hkb_date::{date::SimpleDate, duration::HumanizedDuration};
+use notify_rust::{Notification, Timeout};
 use tokio::{
     io::{AsyncReadExt, AsyncWriteExt},
     net::{unix::SocketAddr, UnixStream},
@@ -32,21 +34,47 @@ async fn main() {
 
     info!("Listening: {}", server.get_addr().to_str().unwrap());
 
+    let mut already_reminded: HashSet<i64> = HashSet::new();
     // spawn thread for reminders
-    tokio::spawn(async {
+    tokio::spawn(async move {
         loop {
             debug!(target: "DAEMON", "Checking reminders to notify!");
 
-            let reminders = fetch_reminders(Some(vec![FetchRemindersOption::RemindAtBetween {
-                start_date: SimpleDate::local(),
-                end_date: SimpleDate::local()
-                    .add_duration(hkb_date::date::Duration::Minute(15))
-                    .unwrap(),
-            }]))
-            .unwrap_or(vec![]);
+            let mut reminders =
+                fetch_reminders(Some(vec![FetchRemindersOption::RemindAtBetween {
+                    start_date: SimpleDate::local(),
+                    end_date: SimpleDate::local()
+                        .add_duration(hkb_date::duration::Duration::Minute(15))
+                        .unwrap(),
+                }]))
+                .unwrap_or(vec![]);
+
+            // TODO: support filtering out ids in reminders service
+            reminders = reminders
+                .into_iter()
+                .filter(|reminder| !already_reminded.contains(&reminder.id))
+                .collect();
 
             debug!(target: "DAEMON", "Found {} reminders to notify!", reminders.len());
-            debug!("{:?}", reminders);
+
+            for reminder in reminders {
+                debug!(target: "DAEMON", "Reminder at: {} - current time: {}", reminder.remind_at.to_string(), SimpleDate::local().to_string());
+
+                Notification::new()
+                    .summary(
+                        format!(
+                            "You have a reminder in: {}",
+                            (reminder.remind_at - SimpleDate::local()).to_human_string()
+                        )
+                        .as_str(),
+                    )
+                    .body(reminder.note.as_str())
+                    .auto_icon()
+                    .timeout(Timeout::Milliseconds(3000))
+                    .show()
+                    .unwrap();
+                already_reminded.insert(reminder.id);
+            }
 
             std::thread::sleep(Duration::from_secs(5));
         }
