+++ hkb_date/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod date;
+pub mod duration;
 mod human_date_parsing;
 
 pub use human_date_parsing::HumanDateParser;
